grpc-in-python/requirements.txt
# Dependencies

google-api-python-client~=2.149
grpcio-reflection~=1.67
grpcio-tools~=1.67
grpcio~=1.67
protobuf~=5.28.3
pytest~=8.3.3